+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- [25-7-10] Improved checkpartial contrast if not using dark mode.
 - [25-7-3] Fixed CAT online indicator.
 - [25-6-21] Remove range check on vfo display.
 - [25-6-20] Added a Rotator tab in the configuration dialog.